+++ R/build.r
@@ -120,7 +120,7 @@
 #' * News (if present).
 #' * An icon linking to the source repository (currently only GitHub and GitLab are supported)
 #'
-#' You can override (but not remove) these defaults with the  `navbar` field.
+#' You can override (and even remove) these defaults with the  `navbar` field.
 #' It has two primary
 #' components: `structure` and `components`. These components interact in
 #' a somewhat complicated way, but the complexity allows you to make minor
@@ -130,8 +130,7 @@
 #' The `structure` defines the layout of the navbar, i.e. the order
 #' of the components, and whether they're right aligned or left aligned.
 #' You can use this component to change the order of the default components,
-#' and to add your own components.
-#' Any unplaced components go to the right of the left navbar.
+#' remove some default components and add your own components.
 #'
 #' ```
 #' navbar: